Add new features to master

+++ src/main/java/com/example/service/impl/UserServiceImpl.java
@@ -23,12 +23,12 @@ import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.NoResultException;
-import java.io.File;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
-import java.util.UUID;
 import java.util.stream.Collectors;
+
+import static com.example.utils.EncodingUtils.encode;
 
 @Service
 @RequiredArgsConstructor
@@ -41,8 +41,6 @@ public class UserServiceImpl implements UserService {
     private final UserTaskDao userTaskDao;
 
     private final ModelMapper modelMapper;
-
-    private final EncodingService encodingService;
 
     @Override
     public MainUserDto getUserById(int id) {
@@ -59,7 +57,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(int id, ApiKeyDto apiKeyDto) {
-        id = getByApiKey(apiKeyDto.getApiKey());
+        int apiKeyId = getByApiKey(apiKeyDto.getApiKey());
+        if (id != apiKeyId) {
+            throw new BadCredentialsException("Your apiKey is not tied to this id");
+        }
         User user = getById(id);
         userDao.delete(user);
     }
@@ -68,11 +69,7 @@ public class UserServiceImpl implements UserService {
     public MainUserDto updateUser(UpdateUserDto userDto) {
         int id = getByApiKey(userDto.getApiKey());
         User oldUser = getById(id);
-        String photo = userDto.getPhoto();
-        String fileName = UUID.randomUUID().toString();
-        File photoFile = new File("Unity/src/main/resources/static/image/" + fileName + ".jpg");
-        encodingService.decodeImage(photo, photoFile.getAbsolutePath());
-        oldUser.setPhoto(fileName);
+        oldUser.setPhoto(userDto.getPhoto());
         oldUser.setFirstName(userDto.getFirstName());
         oldUser.setLastName(userDto.getLastName());
         oldUser.setAboutUser(userDto.getAboutUser());
@@ -89,7 +86,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public MainTaskUserDto takePartInTask(int userId, int taskId, UserTaskDto userTaskDto) {
-        userId = getByApiKey(userTaskDto.getApiKey());
+        int apiKeyId = getByApiKey(userTaskDto.getApiKey());
+        if (userId != apiKeyId) {
+            throw new BadCredentialsException("Your apiKey is not tied to this id");
+        }
         UserTask userTask = modelMapper.map(userTaskDto, UserTask.class);
         User participant = getById(userId);
         Task participatedTask = getTaskById(taskId);
@@ -97,7 +97,7 @@ public class UserServiceImpl implements UserService {
         userTask.setTask(participatedTask);
         userTask.setParticipationDate(LocalDate.now());
 
-        if (!userTask.getTask().isActive()) {
+        if (userTask.getTask().getStatus().getTaskStatus().equalsIgnoreCase("done")) {
             throw new TaskIsNotActiveException("Task is not active.");
         }
 
@@ -128,14 +128,14 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public MainUserTaskDto approveUserForTask(int userId, int taskId, boolean approved, ApiKeyDto apiKeyDto) {
-        int id = getByApiKey(apiKeyDto.getApiKey());
+        int creatorId = getByApiKey(apiKeyDto.getApiKey());
         Task task = getTaskById(taskId);
 
         if (task.getStatus() == Status.DONE) {
             throw new TaskDoneException("Task already done");
         }
 
-        if (task.getCreator().getId() != id) {
+        if (task.getCreator().getId() != creatorId) {
             throw new BadCredentialsException("This user can not approve");
         }
 
@@ -181,7 +181,6 @@ public class UserServiceImpl implements UserService {
     }
 
     private UserTask getByUserIdAndTaskId(int userId, int taskId) {
-
         UserTask userTask;
         try {
             userTask = userTaskDao.getByUserAndTask(userId, taskId);
@@ -211,7 +210,7 @@ public class UserServiceImpl implements UserService {
     public int getByApiKey(String apiKey) {
         User user;
         try {
-            user = userDao.getByApiKey(encodingService.encode(apiKey));
+            user = userDao.getByApiKey(encode(apiKey));
         } catch (NoResultException | EmptyResultDataAccessException exception) {
             throw new BadCredentialsException("API key is invalid");
         }
